refactor daemon thread management and loop processing in zitilib
allow for multiple `Ziti_lib_init()/Ziti_lib_shutdown()` cycles
add checks to prevent hangups when API is called without `Ziti_lib_init()` called first
