+++ library/zitilib/zl.h
@@ -1,16 +1,18 @@
 // Copyright (c) 2026.  NetFoundry Inc
 //
-// 	Licensed under the Apache License, Version 2.0 (the "License");
-// 	you may not use this file except in compliance with the License.
-// 	You may obtain a copy of the License at
+// SPDX-License-Identifier: Apache-2.0
 //
-// 	https://www.apache.org/licenses/LICENSE-2.0
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
 //
-// 	Unless required by applicable law or agreed to in writing, software
-// 	distributed under the License is distributed on an "AS IS" BASIS,
-// 	WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// 	See the License for the specific language governing permissions and
-// 	limitations under the License.
+// https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
 
 //
 //
@@ -89,6 +91,7 @@ int connect_socket(int af, ziti_socket_t clt_sock, ziti_socket_t *ziti_sock);
 bool zl_is_blocking(ziti_socket_t s);
 int zl_socket_af(ziti_socket_t s);
 void zl_set_error(int err);
+bool zl_check_daemon();
 
 ZITI_FUNC
 const char *Ziti_lookup(in_addr_t addr);
